new-project/pdf_auto_classifier/requirements.txt
watchdog==3.0.0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
loguru==0.7.2
PyYAML==6.0.1
requests==2.31.0
pathlib==1.0.1